Add files via upload
Color image quantization is a process of selecting a set of colors to display an image with some representative colors, without a noticeable perceived difference. This algorithm deals with the problem of the quantization errors evaluation by taking into account a probabilistic measure of color space. Developed an ImageJ plugin to minimize the possible degradation of the color profile.